Fix 0 decimal point in second operand
As is, selecting the decimal point after selecting an operator correctly
creates an operand that starts with "0." (0 and a decimal point). Howev-
er, this new number gets incorrectly assigned to the first operand, in
place of what was before selecting the operator, thus getting into a sta-
te that expects a second operand for a new operation. Fix that by commit-
ting the operator when selecting a decimal point after it.